+++ addons/account/__manifest__.py
@@ -103,6 +103,7 @@ You could use this simplified accounting in case you work with an (external) acc
             'account/static/src/js/tax_totals.js',
             'account/static/src/js/legacy_tax_totals.js',
             'account/static/src/js/section_and_note_fields_backend.js',
+            'account/static/src/js/legacy_section_and_note_fields_backend.js',
             'account/static/src/js/tours/account.js',
             'account/static/src/js/bills_upload.js',
             'account/static/src/js/account_selection.js',
